+++ packages/library-base/src/commonMain/kotlin/io/realm/RealmObject.kt
@@ -20,6 +20,10 @@ import io.realm.internal.MutableRealmImpl
 import io.realm.internal.RealmObjectInternal
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.realmObjectInternal
+import io.realm.notifications.DeletedObject
+import io.realm.notifications.InitialObject
+import io.realm.notifications.ObjectChange
+import io.realm.notifications.UpdatedObject
 import kotlinx.coroutines.flow.Flow
 
 /**
@@ -71,6 +75,26 @@ public fun RealmObject.isManaged(): Boolean {
 }
 
 /**
+ * Checks whether [this] and [other] represent the same underlying object or not. It allows to check
+ * if two object from different frozen realms share their object key, and thus represent the same
+ * object at different points in time (= at two different frozen realm versions).
+ */
+internal fun RealmObject.hasSameObjectKey(other: RealmObject?): Boolean {
+    if ((other == null) || (other !is RealmObjectInternal)) return false
+
+    if (!isManaged() || !other.isManaged()) {
+        throw IllegalStateException("Cannot compare unmanaged objects.")
+    }
+
+    val thisKey =
+        RealmInterop.realm_object_get_key(this.realmObjectInternal().`$realm$ObjectPointer`!!)
+    val otherKey =
+        RealmInterop.realm_object_get_key(other.realmObjectInternal().`$realm$ObjectPointer`!!)
+
+    return thisKey == otherKey
+}
+
+/**
  * Returns true if this object is still valid to use, i.e. the Realm is open and the underlying object has
  * not been deleted. Unmanaged objects are always valid.
  */
@@ -90,19 +114,20 @@ public fun RealmObject.isValid(): Boolean {
 }
 
 /**
- * Observe changes to a Realm object. Any change to the object, will cause the flow to emit the updated
- * object. If the observed object is deleted from the Realm, the flow will complete, otherwise it will
- * continue running until canceled.
+ * Observe changes to a Realm object. The flow would emit an [InitialObject] once subscribed and
+ * then, on every change to the object an [UpdatedObject]. If the observed object is deleted from
+ * the Realm, the flow would emit a [DeletedObject] and then will complete, otherwise it will continue
+ * running until canceled.
  *
  * The change calculations will on on the thread represented by [Configuration.notificationDispatcher].
  *
  * @return a flow representing changes to the object.
  */
-public fun <T : RealmObject> T.observe(): Flow<T> {
+public fun <T : RealmObject, C : ObjectChange<T>> T.asFlow(): Flow<ObjectChange<T>> {
     checkNotificationsAvailable()
     val internalObject = this as RealmObjectInternal
     @Suppress("UNCHECKED_CAST")
-    return (internalObject.`$realm$Owner`!!).owner.registerObserver(this) as Flow<T>
+    return (internalObject.`$realm$Owner`!!).owner.registerObserver(this) as Flow<ObjectChange<T>>
 }
 
 private fun RealmObject.checkNotificationsAvailable() {