+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmImpl.kt
@@ -1,7 +1,21 @@
+/*
+ * Copyright 2021 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.realm.internal
 
-import io.realm.Callback
-import io.realm.Cancellable
 import io.realm.MutableRealm
 import io.realm.Realm
 import io.realm.RealmObject
@@ -10,6 +24,9 @@ import io.realm.internal.interop.RealmCoreException
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.platform.runBlocking
 import io.realm.internal.schema.RealmSchemaImpl
+import io.realm.notifications.RealmChange
+import io.realm.notifications.internal.InitialRealmImpl
+import io.realm.notifications.internal.UpdatedRealmImpl
 import io.realm.query.RealmQuery
 import kotlinx.atomicfu.AtomicRef
 import kotlinx.atomicfu.atomic
@@ -19,6 +36,10 @@ import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.asSharedFlow
+import kotlinx.coroutines.flow.flattenConcat
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.takeWhile
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
@@ -28,14 +49,14 @@ import kotlin.reflect.KClass
 internal class RealmImpl private constructor(
     configuration: InternalConfiguration,
     dbPointer: NativePointer
-) : BaseRealmImpl(configuration), Realm {
+) : BaseRealmImpl(configuration), Realm, Flowable<RealmChange<Realm>> {
 
     private val realmPointerMutex = Mutex()
 
     internal val realmScope =
         CoroutineScope(SupervisorJob() + configuration.notificationDispatcher)
     private val realmFlow =
-        MutableSharedFlow<RealmImpl>(replay = 1) // Realm notifications emit their initial state when subscribed to
+        MutableSharedFlow<RealmChange<Realm>>() // Realm notifications emit their initial state when subscribed to
     private val notifier =
         SuspendableNotifier(this, configuration.notificationDispatcher)
     internal val writer =
@@ -69,7 +90,6 @@ internal class RealmImpl private constructor(
         realmReference = frozenReference
         // Update the Realm if another process or the Sync Client updates the Realm
         realmScope.launch {
-            realmFlow.emit(this@RealmImpl)
             notifier.realmChanged().collect { realmReference ->
                 updateRealmPointer(realmReference)
             }
@@ -128,34 +148,15 @@ internal class RealmImpl private constructor(
         }
     }
 
-    override fun observe(): Flow<RealmImpl> {
-        return realmFlow.asSharedFlow()
+    override fun asFlow(): Flow<RealmChange<Realm>> {
+        return flowOf(
+            flow { emit(InitialRealmImpl(this@RealmImpl)) },
+            realmFlow.asSharedFlow().takeWhile { !isClosed() }
+        ).flattenConcat()
     }
 
-    /**
-     * FIXME Hidden until we can add proper support
-     */
-    internal fun addChangeListener(): Cancellable {
-        TODO()
-    }
-
-    override fun <T> registerObserver(t: Thawable<T>): Flow<T> {
+    override fun <T, C> registerObserver(t: Thawable<Observable<T, C>>): Flow<C> {
         return notifier.registerObserver(t)
-    }
-
-    internal override fun <T : RealmObject> registerResultsChangeListener(
-        results: RealmResultsImpl<T>,
-        callback: Callback<RealmResultsImpl<T>>
-    ): Cancellable {
-        TODO("Not yet implemented")
-    }
-
-    internal override fun <T : RealmObject> registerListChangeListener(list: List<T>, callback: Callback<List<T>>): Cancellable {
-        TODO("Not yet implemented")
-    }
-
-    internal override fun <T : RealmObject> registerObjectChangeListener(obj: T, callback: Callback<T?>): Cancellable {
-        TODO("Not yet implemented")
     }
 
     private suspend fun updateRealmPointer(newRealmReference: FrozenRealmReference) {
@@ -173,7 +174,7 @@ internal class RealmImpl private constructor(
                 newRealmReference
             }
             // Notify public observers that the Realm changed
-            realmFlow.emit(this)
+            realmFlow.emit(UpdatedRealmImpl(this))
         }
     }
 
@@ -190,6 +191,7 @@ internal class RealmImpl private constructor(
                 // The local realmReference is pointing to a realm reference managed by either the
                 // version tracker, writer or notifier, so it is already closed
                 super.close()
+                realmFlow.emit(UpdatedRealmImpl(this@RealmImpl))
             }
         }
         // TODO There is currently nothing that tears down the dispatcher