+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmListInternal.kt
@@ -25,6 +25,10 @@ import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.RealmCoreException
 import io.realm.internal.interop.RealmInterop
 import io.realm.internal.interop.Timestamp
+import io.realm.notifications.ListChange
+import io.realm.notifications.internal.DeletedListImpl
+import io.realm.notifications.internal.InitialListImpl
+import io.realm.notifications.internal.UpdatedListImpl
 import kotlinx.coroutines.channels.ChannelResult
 import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.flow.Flow
@@ -34,7 +38,7 @@ import kotlin.reflect.KClass
  * Implementation for unmanaged lists, backed by a [MutableList].
  */
 internal class UnmanagedRealmList<E> : RealmList<E>, MutableList<E> by mutableListOf() {
-    override fun asFlow(): Flow<RealmList<E>> =
+    override fun asFlow(): Flow<ListChange<E>> =
         throw UnsupportedOperationException("Unmanaged lists cannot be observed.")
 }
 
@@ -44,7 +48,7 @@ internal class UnmanagedRealmList<E> : RealmList<E>, MutableList<E> by mutableLi
 internal class ManagedRealmList<E>(
     private val nativePointer: NativePointer,
     private val metadata: ListOperatorMetadata
-) : AbstractMutableList<E>(), RealmList<E>, Observable<ManagedRealmList<E>> {
+) : AbstractMutableList<E>(), RealmList<E>, Observable<ManagedRealmList<E>, ListChange<E>>, Flowable<ListChange<E>> {
 
     private val operator = ListOperator<E>(metadata)
 
@@ -117,7 +121,7 @@ internal class ManagedRealmList<E>(
         }
     }
 
-    override fun asFlow(): Flow<ManagedRealmList<E>> {
+    override fun asFlow(): Flow<ListChange<E>> {
         metadata.realm.checkClosed()
         return metadata.realm.owner.registerObserver(this)
     }
@@ -141,14 +145,22 @@ internal class ManagedRealmList<E>(
     override fun emitFrozenUpdate(
         frozenRealm: RealmReference,
         change: NativePointer,
-        channel: SendChannel<ManagedRealmList<E>>
+        channel: SendChannel<ListChange<E>>
     ): ChannelResult<Unit>? {
         val frozenList: ManagedRealmList<E>? = freeze(frozenRealm)
         return if (frozenList != null) {
-            channel.trySend(frozenList)
+            val builder = ListChangeSetBuilderImpl(change)
+
+            if (builder.isEmpty()) {
+                channel.trySend(InitialListImpl(frozenList))
+            } else {
+                channel.trySend(UpdatedListImpl(frozenList, builder.build()))
+            }
         } else {
-            channel.close()
-            null
+            channel.trySend(DeletedListImpl(UnmanagedRealmList()))
+                .also {
+                    channel.close()
+                }
         }
     }