+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmResultsImpl.kt
@@ -22,6 +22,9 @@ import io.realm.internal.interop.Callback
 import io.realm.internal.interop.NativePointer
 import io.realm.internal.interop.RealmCoreException
 import io.realm.internal.interop.RealmInterop
+import io.realm.notifications.ResultsChange
+import io.realm.notifications.internal.InitialResultsImpl
+import io.realm.notifications.internal.UpdatedResultsImpl
 import kotlinx.coroutines.channels.ChannelResult
 import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.flow.Flow
@@ -39,7 +42,7 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
     private val clazz: KClass<E>,
     private val mediator: Mediator,
     private val mode: Mode = Mode.RESULTS
-) : AbstractList<E>(), RealmResults<E>, Observable<RealmResultsImpl<E>>, RealmStateHolder {
+) : AbstractList<E>(), RealmResults<E>, Observable<RealmResultsImpl<E>, ResultsChange<E>>, RealmStateHolder, Flowable<ResultsChange<E>> {
 
     enum class Mode {
         // FIXME Needed to make working with @LinkingObjects easier.
@@ -70,7 +73,7 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
         }
     }
 
-    override fun asFlow(): Flow<RealmResultsImpl<E>> {
+    override fun asFlow(): Flow<ResultsChange<E>> {
         realm.checkClosed()
         return realm.owner.registerObserver(this)
     }
@@ -108,10 +111,17 @@ internal class RealmResultsImpl<E : RealmObject> constructor(
     override fun emitFrozenUpdate(
         frozenRealm: RealmReference,
         change: NativePointer,
-        channel: SendChannel<RealmResultsImpl<E>>
+        channel: SendChannel<ResultsChange<E>>
     ): ChannelResult<Unit>? {
         val frozenResult = freeze(frozenRealm)
-        return channel.trySend(frozenResult)
+
+        val builder = ListChangeSetBuilderImpl(change)
+
+        return if (builder.isEmpty()) {
+            channel.trySend(InitialResultsImpl(frozenResult))
+        } else {
+            channel.trySend(UpdatedResultsImpl(frozenResult, builder.build()))
+        }
     }
 
     override fun realmState(): RealmState = realm